DocumentSymbol uses SymbolInformation
When flat list of information is calculated by the LSP
Server it needs to use SymbolInformation.

This commit makes use of the 'namesResolvedAs', which
returns all the names from the active context.
Although the generated list of symbols still needs a lot
of improvements, as it should generate a tree of the AST,
rather just blindly enumerate the names.